refactor dataset specs
